Move AppInput to mobile molecules and update component exports
