--- apps/mobile/src/components/atoms/AppButton/index.ts
@@ -1,6 +1,0 @@
-export type {
-  AppButtonProps,
-  AppButtonSize,
-  AppButtonVariant,
-} from './AppButton';
-export { AppButton } from './AppButton';
