--- apps/mobile/src/components/atoms/AppInput/index.ts
@@ -1,2 +1,0 @@
-export type { AppInputProps } from './AppInput';
-export { AppInput } from './AppInput';
