--- apps/mobile/src/components/atoms/AppLabel/index.ts
@@ -1,2 +1,0 @@
-export type { AppLabelProps } from './AppLabel';
-export { AppLabel } from './AppLabel';
