--- apps/mobile/src/components/atoms/AppText/index.ts
@@ -1,2 +1,0 @@
-export type { AppTextProps, AppTextVariant, AppTextWeight } from './AppText';
-export { AppText } from './AppText';
