--- apps/mobile/src/components/atoms/Surface/index.ts
@@ -1,2 +1,0 @@
-export type { SurfaceProps } from './Surface';
-export { Surface } from './Surface';
