+++ apps/mobile/src/components/atoms/index.ts
@@ -1,5 +1,4 @@
-export * from './AppButton';
-export * from './AppInput';
-export * from './AppLabel';
-export * from './AppText';
-export * from './Surface';
+export * from './AppButton/AppButton';
+export * from './AppLabel/AppLabel';
+export * from './AppText/AppText';
+export * from './Surface/Surface';
