+++ apps/mobile/src/components/molecules/AppInput/AppInput.spec.tsx
