+++ apps/mobile/src/components/molecules/index.ts
@@ -1,2 +1,3 @@
 // Molecules are combinations of atoms forming a distinct UI unit.
 // Examples: FormField (Label + Input), TodoItem (Checkbox + Label)
+export * from './AppInput/AppInput';
